handle-out: sign convert

+++ model/bgp-ns3-socket-out.cc
@@ -7,7 +7,7 @@ BgpNs3SocketOut::BgpNs3SocketOut(Ptr<Socket> socket) {
 }
 
 bool BgpNs3SocketOut::handleOut(const uint8_t *buffer, size_t length) {
-    return _socket->Send(buffer, length, 0) == length;
+    return _socket->Send(buffer, length, 0) == (int) length;
 }
 
 }
